+++ web_thing_server/include/web_thing_property.h
@@ -22,7 +22,7 @@ typedef struct enum_item_t enum_item_t;
 /*
  * if new value is different then the old one set_callback_t must return "1" (one)
  */
-typedef int8_t (set_callback_t)(char *new_value);
+typedef int16_t (set_callback_t)(char *new_value);
 
 /*
  * function jsonize_t must return json representation of the property value in
